refactor: :recycle: Adding typescript :fire:
