+++ package.json
@@ -21,6 +21,7 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
-    "vite": "^4.4.5"
+    "typescript": "^5.4.5",
+    "vite": "^5.2.10"
   }
 }
